Organized MemoryWatcher-related code into a separate class.
